Revert "fixed issues regarding success rates"
This reverts commit 1141aea60764f06a3215c1297a0fe2fc07e5debc.

+++ Source/VCFarming/Public/FarmLand.h
@@ -20,7 +20,7 @@ public:
 	UFUNCTION(BlueprintCallable)
 	void InputCropType(int index, int SuccessRate);
 
-	float CurrentSuccessRate;
+	int CurrentSuccessRate;
 protected:
 	// Called when the game starts or when spawned
 	virtual void BeginPlay() override;
