+++ Source/VCFarming/Public/FarmLand2.h
@@ -32,7 +32,7 @@ public:
 
 	void InputCropType(int index, int SuccessRate);
 
-	float CurrentSuccessRate;
+	int CurrentSuccessRate;
 
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
